order_processing_and_subscription_service/requirements.txt
Django==3.2.25
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-telegram-bot==20.7
SQLAlchemy==1.4.52
requests==2.31.0
python-dotenv==1.0.0
